Moved Node and Linkedlist class to one source file & tested addToFront, addToEnd, SearchList methods of linkedlist class
